// invest-calc-app/src/common/types.ts
export enum InvestmentOptionEnum {
    NOT_SELECTED = 'NOT_SELECTED',
    CASH_INVESTMENTS = 'CASH_INVESTMENTS',
    FIXED_INTEREST = 'FIXED_INTEREST',
    SHARES = 'SHARES',
    MANAGED_FUNDS = 'MANAGED_FUNDS',
    ETF = 'ETF',
    INVESTMENT_BONDS = 'INVESTMENT_BONDS',
    ANNUITIES = 'ANNUITIES',
    LICs = 'LICs',
    REITs = 'REITs',
}